+++ tasks/build-phaser.js
@@ -20,49 +20,9 @@ var modulesWithDependencies = moduleNames
 
 module.exports = function (grunt) {
 
-    //  Validate module exclusions.
-    function validateExcludes (excludes) {
-        return {
-            //  Fail a validation if conditions are not met.
-            fail: function (filter, reporter) {
-                var result = filter(excludes);
+    var validateExcludes = require('./helpers/validateExcludes')(grunt);
 
-                if (result.length > 0)
-                {
-                    reporter(result);
-
-                    grunt.fail.fatal('Aborting due to invalid parameter input.');
-                }
-
-                return this;
-            },
-
-            //  If a module's dependencies are missing, select that to be
-            //  removed from the build.
-            push: function (filter, reporter) {
-                var result = filter(excludes);
-
-                if (result.length > 0)
-                {
-                    reporter(result);
-                }
-
-                excludes = excludes.concat(result);
-
-                return this;
-            },
-
-            //  Report the excluded modules.
-            result (reporter) {
-                if (excludes.length > 0)
-                {
-                    reporter(excludes);
-                }
-
-                return excludes;
-            }
-        };
-    };
+    var selectBuildFiles = require('./helpers/selectBuildFiles')(modules);
 
     //  Filter required module names given by the user.
     function filterNotOptionalModules (excludes) {
@@ -129,26 +89,6 @@ module.exports = function (grunt) {
         excludes.forEach(function (exclude) {
             grunt.log.writeln("* " + exclude + ' - ' + modules[exclude].description);
         });
-    }
-
-    // Select the module files to be compiled.
-    function selectBuildFiles (excludes) {
-        return moduleNames.reduce(function (filelist, name) {
-            var m = modules[name];
-
-            if (!m.optional || excludes.indexOf(name) < 0)
-            {
-                //  A module is required or is optional but was not excluded.
-                filelist.push(m.files);
-            }
-            else if (excludes.indexOf(name) >= 0 && m.stubs)
-            {
-                //  A module is excluded but has a stub.
-                filelist.push(m.stubs);
-            }
-
-            return filelist;
-        }, []);
     }
